+++ packages/flutter_solidart/example/lib/pages/derived_signal.dart
@@ -9,9 +9,8 @@ class DerivedSignalsPage extends StatefulWidget {
 }
 
 class _DerivedSignalsPageState extends State<DerivedSignalsPage> {
-  late final count = Signal(0, options: SignalOptions(name: 'count'));
-  late final doubleCount =
-      Computed(() => count() * 2, options: SignalOptions(name: 'doubleCount'));
+  late final count = Signal(0, name: 'count');
+  late final doubleCount = Computed(() => count() * 2, name: 'doubleCount');
 
   @override
   Widget build(BuildContext context) {