+++ packages/flutter_solidart/example/lib/pages/resource.dart
@@ -12,12 +12,9 @@ class ResourcePage extends StatefulWidget {
 }
 
 class _ResourcePageState extends State<ResourcePage> {
-  final userId = Signal(1, options: SignalOptions(name: 'userId'));
+  final userId = Signal(1, name: 'userId');
 
-  late final user = Resource(
-      fetcher: fetchUser,
-      source: userId,
-      options: ResourceOptions(name: 'user'));
+  late final user = Resource(fetcher: fetchUser, source: userId, name: 'user');
 
   late final userHairColor = user.select((data) {
     return jsonDecode(data)['hair_color'] as String;