+++ packages/solidart/lib/src/core/core.dart
@@ -22,5 +22,4 @@ part 'read_signal.dart';
 part 'resource.dart';
 part 'signal.dart';
 part 'signal_base.dart';
-part 'signal_options.dart';
 part 'batch.dart';